+++ nasl/nasl_packet_forgery_v6.c
@@ -125,7 +125,7 @@ int n;
  * @return Forged IP packet.
  */
 tree_cell *
-forge_ipv6_packet (lex_ctxt *lexic)
+forge_ip_v6_packet (lex_ctxt *lexic)
 {
   tree_cell *retc;
   struct ip6_hdr *pkt;
@@ -189,52 +189,62 @@ forge_ipv6_packet (lex_ctxt *lexic)
 /**
  * @brief Obtain IPv6 header element.
  *
- * @param[in] lexic   Lexical context of NASL interpreter.
- * @param[in] ipv6    IP v6 header
- * @param[in] element Element to extract from the header.
+ * @param[in] lexic      Lexical context of NASL interpreter.
+ * @param[in] ipv6       IPv6 header. TODO: Once versions older than 20.08 are
+ * no longer in use the parameter name can be changed to 'ip6'.
+ * @param[in] element    Element to extract from the header.
  *
  * @return tree_cell with the IP header element.
  */
 tree_cell *
-get_ipv6_element (lex_ctxt *lexic)
+get_ip_v6_element (lex_ctxt *lexic)
 {
   tree_cell *retc;
-  struct ip6_hdr *ip6 = (struct ip6_hdr *) get_str_var_by_name (lexic, "ipv6");
   char *element = get_str_var_by_name (lexic, "element");
   char ret_ascii[INET6_ADDRSTRLEN];
   int ret_int = 0;
   int flag = 0;
+  struct ip6_hdr *ip6;
 
+  /* Parameter name 'ipv6' was renamed to 'ip6' for consistency.
+   * For backwards compatibility reasons we still need to consider the 'ipv6'
+   * argument.
+   */
+  ip6 = (struct ip6_hdr *) get_str_var_by_name (lexic, "ipv6");
   if (ip6 == NULL)
     {
-      nasl_perror (lexic, "get_ipv6_element: no valid 'ipv6' argument\n");
-      return NULL;
+      ip6 = (struct ip6_hdr *) get_str_var_by_name (lexic, "ip6");
+      if (ip6 == NULL)
+        {
+          nasl_perror (lexic, "%s: no valid 'ip6' argument\n", __func__);
+          return NULL;
+        }
     }
 
   if (element == NULL)
     {
-      nasl_perror (lexic, "get_ipv6_element: no valid 'element' argument\n");
+      nasl_perror (lexic, "%s: no valid 'element' argument\n", __func__);
       return NULL;
     }
 
   if (!strcmp (element, "ip6_v"))
     {
-      ret_int = (ip6->ip6_flow & 0x3ffff);
+      ret_int = ntohl (ip6->ip6_flow) >> 28;
       flag++;
     }
   else if (!strcmp (element, "ip6_tc"))
     {
-      ret_int = (ip6->ip6_flow >> 20) & 0xff;
+      ret_int = (ntohl (ip6->ip6_flow) >> 20) & 0xff;
       flag++;
     }
   else if (!strcmp (element, "ip6_fl"))
     {
-      ret_int = ip6->ip6_flow >> 28;
+      ret_int = ntohl (ip6->ip6_flow) & 0x3ffff;
       flag++;
     }
   else if (!strcmp (element, "ip6_plen"))
     {
-      ret_int = (ip6->ip6_plen);
+      ret_int = UNFIX (ip6->ip6_plen);
       flag++;
     }
   else if (!strcmp (element, "ip6_nxt"))
@@ -292,7 +302,7 @@ get_ipv6_element (lex_ctxt *lexic)
  * @return tree_cell with the forged IP packet.
  */
 tree_cell *
-set_ipv6_elements (lex_ctxt *lexic)
+set_ip_v6_elements (lex_ctxt *lexic)
 {
   struct ip6_hdr *o_pkt = (struct ip6_hdr *) get_str_var_by_name (lexic, "ip6");
   int size = get_var_size_by_name (lexic, "ip6");
@@ -302,7 +312,7 @@ set_ipv6_elements (lex_ctxt *lexic)
 
   if (o_pkt == NULL)
     {
-      nasl_perror (lexic, "set_ipv6_elements: missing <ip6> field\n");
+      nasl_perror (lexic, "%s: missing <ip6> field\n", __func__);
       return NULL;
     }
 
@@ -333,7 +343,7 @@ set_ipv6_elements (lex_ctxt *lexic)
  * @return Print and returns FAKE_CELL.
  */
 tree_cell *
-dump_ipv6_packet (lex_ctxt *lexic)
+dump_ip_v6_packet (lex_ctxt *lexic)
 {
   int i;
   char addr[INET6_ADDRSTRLEN];
@@ -391,7 +401,7 @@ dump_ipv6_packet (lex_ctxt *lexic)
  * @return the modified datagram.
  */
 tree_cell *
-insert_ipv6_options (lex_ctxt *lexic)
+insert_ip_v6_options (lex_ctxt *lexic)
 {
   struct ip6_hdr *ip6 = (struct ip6_hdr *) get_str_var_by_name (lexic, "ip6");
   int code = get_int_var_by_name (lexic, "code", 0);
@@ -410,8 +420,10 @@ insert_ipv6_options (lex_ctxt *lexic)
 
   if (ip6 == NULL)
     {
-      nasl_perror (lexic, "Usage : insert_ipv6_options(ip6:<ip6>, code:<code>, "
-                          "length:<len>, value:<value>\n");
+      nasl_perror (lexic,
+                   "Usage : %s(ip6:<ip6>, code:<code>, "
+                   "length:<len>, value:<value>\n",
+                   __func__);
       return NULL;
     }